requirements.txt
gymnasium>=0.29.1
stable-baselines3>=2.2.1
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
python-dotenv>=1.0.0
matplotlib>=3.7.0
requests>=2.31.0 
streamlit
plotly
schedule>=1.2.0
pytz 